requirements.txt
fastapi
pydantic
sqlalchemy
psycopg
uvicorn
email-validator
redis
minio
argon2-cffi
python-dotenv
urllib3
booktype